Merged PR 7104: This is to fix some broken links related to the un-named structs and some title fixes for CATS tests
This is to fix some broken links related to the un-named structs and some title fixes for CATS tests related to trailing = signs that were being seen as %3 and showing as broken links.
There are still likely issues with un-named structs, but this should reduce some of the noise at least.