Add FLT_FILE_NAME_QUERY_ALWAYS_ALLOW_CACHE_LOOKUP
Add FLT_FILE_NAME_QUERY_ALWAYS_ALLOW_CACHE_LOOKUP entry to the table of NameOptions parameter based on the comment from devs ("the fact that FLT_FILE_NAME_QUERY_ALWAYS_ALLOW_CACHE_LOOKUP is not documented for FltGetFileNameInformationUnsafe() appears to be an oversight.  There’s nothing in Filter Manager that would prevent using that flag with that API").