Merged PR 15531: This is to attempt to add the correct unnamed struct and enum number so that Meriwether will publish these ref topics.
This is to attempt to add the correct unnamed struct and enum number so that Meriwether will publish these ref topics.

Because the review site is pinned to the next release branch it is difficult to get these right.

KSAUDIO_MIXLEVEL topic __unnamed_struct_44
https://review.docs.microsoft.com/en-us/windows-hardware/drivers/ddi/content/ksmedia/ns-ksmedia--ksaudio_mixlevel.md?branch=audio-far-field

INTERLEAVED_AUDIO_FORMAT_INFORMATION Struct 38
 ns-ksmedia-_interleaved_audio_format_information.md /ns-ksmedia-__unnamed_struct_38.md
https://review.docs.microsoft.com/en-us/windows-hardware/drivers/ddi/content/ksmedia/ns-ksmedia-_interleaved_audio_format_information.md?branch=audio-far-field

KSPROPERTY_INTERLEAVEDAUDIO enum 17
https://review.docs.microsoft.com/en-us/windows-hardware/drivers/ddi/content/ksmedia/ne-ksmedia-ksproperty_interleavedaudio.md?branch=audio-far-field