Adding Missing Return Value
The following function documentation had a missing return value. Through my testing and research, the return type of `HIDP_STATUS_USAGE_NOT_FOUND` was missing from this page. I do not actually believe I got all the return types though as there very well might be more. Sadly, the header file did not have a function header for this function. But, bridging the gap a little bit more is useful!

I have also added in the actual NTSTATUS codes too, because Visual Studio, for example, only shows the hexadecimal values, unlike when it sees a HRESULT. This is also just helpful in general to tie a number to a human readable string.