Add pipes to the left and right of table data
The documentation on https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/portcls/nf-portcls-iminiportwavertstream-allocateaudiobuffer does not seem to be rendering that table correctly.

Looking at the table documentation, I see that tables have pipes to the left and right of columns, so I added those additional pipes to see if it renders correctly. Also aligned the whole table in source code to make it easier to read.