Merged PR 16801: Implement Documentation for new CreateCommonBufferFromMdl DMA Function
This change adds documentation for the new CreateCommonBufferFromMdl DMA function. Documentatino is also provided for the new configuration types usable by the new function.

The code change was recently checked in, so stubs currently do not exist for this file. Modifications will be made once the changes have made it into rs_prerelease.