Add note about IRQL restriction for non paging IO.
The SAL for `FltReissueSynchronousIo` states ```
_When_(FlagOn(CallbackData->Iopb->IrpFlags, IRP_PAGING_IO), _IRQL_requires_max_(APC_LEVEL))
_When_(!FlagOn(CallbackData->Iopb->IrpFlags, IRP_PAGING_IO), _IRQL_requires_max_(PASSIVE_LEVEL))
```

(IOW: MUST be at PASSIVE unless for paging when it MAY be issued at APC_LEVEL).

This cumbersome change tries to do this.  But I am abusing the req.irql macro so it may not work - in which case it will need to go into remarks.