Add when this callback is invoked
This document doesn't tell about "when this callback is invoked." It is described in the following document, but there is some difficulty to tie these documents for non English native speakers.

Directed power management
https://docs.microsoft.com/en-us/windows-hardware/design/device-experiences/directed-power-management

DFx
DFx is an extension to PoFx that enables the power manager to direct a device to enter a lower power state. DFx takes effect after a DRIPS constraint device has been active for at least two consecutive minutes (driver-configurable timeout, default 2 minutes) when the system is in Modern Standby and there is no activator-brokered software activity running. DFx directs devices to enter their target D-state for Modern Standby with the ability to arm for wake as they may during a normal Modern Standby transition. It does not carry the assumptions associated with an S-IRP, e.g. device reset is not required. DFx will not power down paging or debug devices.