Specify maximum length of common buffer
It seems that, no surprise, WdfCommonBufferCreate is subject to the dumbness that is the underlying DmaOperations->AllocateCommonBuffer function, for which the length is an ULONG (qv).  This was terrific 20 years ago.  Today, not so much.  But I digress.

ANYhow...  we should probably let people know the max size.

In passing, I also noted that you can delete the Common Buffer object by calling WdfObjectDelete.