Update note about the checksum fallback behavior
A customer relied on the behavior mentioned in this note and expected their miniport driver to pass the HLK test with this behavior. However, the HLK test requires the miniport to validate the checksum on all packets before indicating it up as there may be other protocols on the system may not do the fallback. The fallback mentioned in this note is specific to the TCPIP protocol driver.