Update nf-wdm-kewaitforsingleobject.md
Similarly to KeWaitForMultipleObjects, the dispatcher object passed to KeWaitForSingleObject must be allocated in nonpaged system memory - see https://learn.microsoft.com/en-us/windows-hardware/drivers/ddi/wdm/nf-wdm-kewaitformultipleobjects. This is because KeWaitForSingleObject raises IRQL to DISPATCH_LEVEL before locking the object, so if it is allocated in paged memory a pagefault will lead to System crash.