Do not expect that Acquires will be matched by Release
This is very grotty and edge case but kinda matters (and is also incredibly
non-intuitive unless you think too har about the FltMgr implementation) but
as a good rule you should not acquire any resources in the Acquire calls (which
provoke many philosophical questions).

This is a bad initial attempt to explain this.  I'm sure Lori and ted will do
better.