Widen warning about when not to call FltCreateFile
In particular you should not send any creates down (this change is for FltCreateFile but the same is true for IoCreateFile and all the FltCreateFile variants) to FASTFAT with kernel APCs disabled at all.  If the IRP is cancelled FASTFAT will see an eventual Oplock Break come in as an APC (for some magical reason NTFS is OK).

EnhancedOplocks var 13 subvar 1 is a good repro.