RestartKey has to be supplied, *RestartKey may be null.
*This will need changed to be correct markup*

The documentation says a lot of things like

     If RestartKey is NULL,

When it should say

     If *RestartKey is NULL

or (in a more ugly manner)

     Of the address pointed to by RestartKey is NULL

If you actually to specify the parameter as NULL things go sidesways pretty fast and empirically specifying a restart key which points to NULL does what is described in the doco.

This PR tried to capture that in a couple of places. Its wrong because I dunno how to specify a * in a italicized chunk of code.

I may have missed some places to