Update ne-netadapter-net_packet_filter_flags.md
The current documentation seems to indicate that the filters are optional, but if the driver lacks support for any of the filters, the driver is likely to fail to bind to upper layers, causing confusing and hard-to-diagnose issues.

NDIS docs indicate that set OID_GEN_CURRENT_PACKET_FILTER is mandatory. Drivers must not fail this set -- failure will prevent the upper layer from binding.

NetAdapterCx's handler for this OID will fail the OID if the device does not support all of the requested filter types.

Therefore, it is pretty much mandatory for the driver to report support for all filter types even if they don't actually support them.